pocketurls/aiopocketurls/requirements.txt
aiohttp
aiohttp_jinja2
aiopg
aioredis
gunicorn
pyyaml
trafaret
